Re-factored _caster.py to caster.py in apps
