+++ castervoice/apps/caster.py
@@ -1,0 +1,104 @@
+from castervoice.lib.imports import *
+from castervoice.lib.dfplus.merge.mergepair import MergeInf
+from castervoice.lib.dfplus.merge.mergerule import MergeRule
+from castervoice.lib.ctrl.dependencies import pip_path, update
+
+_NEXUS = control.nexus()
+
+
+class DependencyUpdate(RunCommand):
+    synchronous = True
+
+    # pylint: disable=method-hidden
+    def process_command(self, proc):
+        # Process the output from the command.
+        RunCommand.process_command(self, proc)
+        # Only reboot dragon if the command was successful and online_mode is true
+        # 'pip install ...' may exit successfully even if there were connection errors.
+        if proc.wait() == 0 and update:
+            Playback([(["reboot", "dragon"], 0.0)]).execute()
+
+
+def change_monitor():
+    if settings.SETTINGS["sikuli"]["enabled"]:
+        Playback([(["monitor", "select"], 0.0)]).execute()
+    else:
+        print("This command requires SikuliX to be enabled in the settings file")
+
+
+class CasterRule(MergeRule):
+    @staticmethod
+    def generate_ccr_choices(nexus):
+        choices = {}
+        for ccr_choice in nexus.merger.global_rule_names():
+            choices[ccr_choice] = ccr_choice
+        return Choice("name", choices)
+
+    @staticmethod
+    def generate_sm_ccr_choices(nexus):
+        choices = {}
+        for ccr_choice in nexus.merger.selfmod_rule_names():
+            choices[ccr_choice] = ccr_choice
+        return Choice("name2", choices)
+
+    mapping = {
+        # update management
+        "update caster":
+            R(DependencyUpdate([pip_path, "install", "--upgrade", "castervoice"])),
+        "update dragonfly":
+            R(DependencyUpdate([pip_path, "install", "--upgrade", "dragonfly2"])),
+
+        # hardware management
+        "volume <volume_mode> [<n>]":
+            R(Function(navigation.volume_control, extra={'n', 'volume_mode'})),
+        "change monitor":
+            R(Key("w-p") + Pause("100") + Function(change_monitor)),
+
+        # window management
+        'minimize':
+            R(Playback([(["minimize", "window"], 0.0)])),
+        'maximize':
+            R(Playback([(["maximize", "window"], 0.0)])),
+        "remax":
+            R(Key("a-space/10,r/10,a-space/10,x")),
+
+        # mouse alternatives
+        "legion [<monitor>]":
+            R(Function(navigation.mouse_alternates, mode="legion", nexus=_NEXUS)),
+        "rainbow [<monitor>]":
+            R(Function(navigation.mouse_alternates, mode="rainbow", nexus=_NEXUS)),
+        "douglas [<monitor>]":
+            R(Function(navigation.mouse_alternates, mode="douglas", nexus=_NEXUS)),
+
+        # ccr de/activation
+        "<enable> <name>":
+            R(Function(_NEXUS.merger.global_rule_changer(), save=True)),
+        "<enable> <name2>":
+            R(Function(_NEXUS.merger.selfmod_rule_changer(), save=True)),
+        "enable caster":
+            R(Function(_NEXUS.merger.merge, time=MergeInf.RUN, name="numbers")),
+        "disable caster":
+            R(Function(_NEXUS.merger.ccr_off)),
+    }
+    extras = [
+        IntegerRefST("n", 1, 50),
+        Dictation("text"),
+        Dictation("text2"),
+        Dictation("text3"),
+        Choice("enable", {
+            "enable": True,
+            "disable": False
+        }),
+        Choice("volume_mode", {
+            "mute": "mute",
+            "up": "up",
+            "down": "down"
+        }),
+        generate_ccr_choices.__func__(_NEXUS),
+        generate_sm_ccr_choices.__func__(_NEXUS),
+        IntegerRefST("monitor", 1, 10)
+    ]
+    defaults = {"n": 1, "nnv": 1, "text": "", "volume_mode": "setsysvolume", "enable": -1}
+
+
+control.non_ccr_app_rule(CasterRule(), context=None, rdp=False)
